+++ Source/Common/DynBitSet.h
@@ -10,19 +10,19 @@ public:
     // Ctor; performs zero-initialization.
     DynBitSet();
     // Ctor; takes the size (number of bits) as input.
-    explicit DynBitSet(const uint size);
+    explicit DynBitSet(const size_t size);
     // Resets the values of all bits to 'value' (0 or 1).
     void reset(const bool value);
     // Sets the value of the specified bit to 0.
-    void clearBit(const uint index);
+    void clearBit(const size_t index);
     // Sets the value of the specified bit to 1.
-    void setBit(const uint index);
+    void setBit(const size_t index);
     // Inverts the value of the specified bit.
-    void toggleBit(const uint index);
+    void toggleBit(const size_t index);
     // Returns 'true' if the specified bit is 1, 'false' otherwise.
-    bool testBit(const uint index) const;
+    bool testBit(const size_t index) const;
 private:
-    std::unique_ptr<uint[]> m_bits;
-    uint                    m_bitCount;
-    uint                    m_dwordCount;
+    std::unique_ptr<uint32_t[]> m_bits;
+    uint32_t                    m_bitCount;
+    uint32_t                    m_dwordCount;
 };
